# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales:
    post:
      operationId: createSale
      tags: [sales]
      summary: Record a sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Recorded sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "200":
          description: Existing sale for the same offline request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /dining-locations:
    get:
      operationId: getDiningLocations
      tags: [catalog]
      summary: List active tables and dining spaces
      responses:
        "200":
          description: Active dining locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiningLocation"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/dining-locations:
    post:
      operationId: createDiningLocation
      tags: [admin]
      summary: Create a table or dining space
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiningLocationInput"
      responses:
        "201":
          description: Created dining location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiningLocation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /sales/epoch:
    get:
      operationId: getSalesEpoch
      tags: [sales]
      summary: Current sale generation for safe offline synchronization
      responses:
        "200":
          description: Current generation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesEpoch"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/dining-locations/{id}:
    delete:
      operationId: deleteDiningLocation
      tags: [admin]
      summary: Remove a table or dining space from future orders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Dining location removed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Dining location not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /admin/storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage, admin]
      summary: Request a product image upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadRequest"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /storage/objects/{objectPath}:
    get:
      operationId: getStorageObject
      tags: [storage]
      summary: Serve a product image
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stored object
          content:
            image/jpeg: {}
            image/png: {}
            image/webp: {}
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Object not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    SaleItemInput:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
      required: [productId, quantity]
    SaleInput:
      type: object
      properties:
        tableLabel:
          type: string
          minLength: 1
        paymentMethod:
          type: string
          minLength: 1
        cashTendered:
          type: integer
          minimum: 0
        cashAmount:
          type: integer
          minimum: 0
        waveAmount:
          type: integer
          minimum: 0
        orangeMoneyAmount:
          type: integer
          minimum: 0
        clientRequestId:
          type: string
          format: uuid
        salesEpoch:
          type: integer
          minimum: 0
        isTest:
          type: boolean
        testEpoch:
          type: integer
          minimum: 0
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/SaleItemInput"
      required: [tableLabel, paymentMethod, items]
    SalesEpoch:
      type: object
      properties:
        epoch:
          type: integer
        testEpoch:
          type: integer
      required: [epoch, testEpoch]
    SaleItem:
      type: object
      properties:
        id:
          type: string
        saleId:
          type: string
        productId:
          type: string
        name:
          type: string
        unitPrice:
          type: integer
        quantity:
          type: integer
        lineTotal:
          type: integer
      required: [id, saleId, productId, name, unitPrice, quantity, lineTotal]
    Sale:
      type: object
      properties:
        id:
          type: string
        receiptNumber:
          type: string
        isTest:
          type: boolean
        clerkUserId:
          type: string
        tableLabel:
          type: string
        paymentMethod:
          type: string
        subtotal:
          type: integer
        cashTendered:
          type: integer
        cashAmount:
          type: integer
        waveAmount:
          type: integer
        orangeMoneyAmount:
          type: integer
        changeDue:
          type: integer
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
      required:
        [id, receiptNumber, isTest, clerkUserId, tableLabel, paymentMethod, subtotal, cashTendered, cashAmount, waveAmount, orangeMoneyAmount, changeDue, createdAt, items]
...
    DiningLocation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        kind:
          type: string
          enum: [table, space]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, kind, active, createdAt, updatedAt]
    DiningLocationInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 60
        kind:
          type: string
          enum: [table, space]
      required: [name, kind]
...
    UploadRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        size:
          type: integer
          minimum: 1
          maximum: 5242880
        contentType:
          type: string
          enum: [image/jpeg, image/png, image/webp]
      required: [name, size, contentType]
    UploadResponse:
      type: object
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
      required: [uploadURL, objectPath]
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Administrator role required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"